Implementacion de Carga de Servicios y Peluqueros + Inicio de Peluquero

+++ AcquaDiCane/Models/DetalleDelTurno.cs
@@ -15,6 +15,6 @@ namespace AcquaDiCane.Models
         [ForeignKey("TurnoAsignado")]
         public int TurnoAsignadoId { get; set; }
         public Turno TurnoAsignado { get; set; }
-        public double PrecioServicio { get; set; }
+        public decimal PrecioServicio { get; set; }
     }
 }
